requirements.txt
absl-py==0.9.0
astor==0.7.1
astroid==2.3.3
astropy==4.0
atm==0.2.2
atomicwrites==1.3.0
attrs==19.3.0
autokeras==1.0.2
automl-gs==0.2.1
autopep8==1.5
backcall==0.1.0
baytune==0.2.5
bleach==3.1.4
blinker==1.4
bokeh==2.0.1
boto3==1.10.2
botocore==1.13.2
cachetools==3.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cirq==0.7.0
click==7.1.1
cloudpickle==1.3.0
colorama==0.4.3
confuse==1.0.0
contextlib2==0.5.5
coverage==5.0.4
cryptography==2.8
cycler==0.10.0
cytoolz==0.10.1
dash==0.39.0
dash-core-components==0.44.0
dash-cytoscape==0.1.1
dash-html-components==0.14.0
dash-renderer==0.20.0
dash-table==3.6.0
dash-table-experiments==0.6.0
dask==2.13.0
dask-glm==0.2.0
dask-ml==1.2.0
dataclasses==0.7
deap==1.3.1
decorator==4.4.2
deepgaze==0.1
defusedxml==0.6.0
dill==0.3.1.1
distributed==2.13.0
dlib==19.17.0
docutils==0.14
entrypoints==0.3
Flask==1.1.1
Flask-Compress==1.4.0
Flask-Restless==0.17.0
flask-restless-swagger-2==0.0.3
Flask-SQLAlchemy==2.4.1
fsspec==0.7.1
future==0.17.1
gast==0.2.2
gevent==1.4.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
greenlet==0.4.15
grpcio==1.27.2
h5py==2.10.0
HeapDict==1.0.1
htmlmin==0.1.12
httplib2==0.17.1
hyperopt==0.1.2
idna==2.9
imageio==2.5.0
importlib-metadata==1.6.0
interpret==0.1.13
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
json-tricks==3.14.0
json5==0.9.0
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterlab==1.1.4
jupyterlab-server==1.0.7
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tuner==1.0.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lime==0.1.1.36
llvmlite==0.31.0
locket==0.2.0
lockfile==0.12.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mccabe==0.6.1
mimerender==0.6.0
missingno==0.4.2
mistune==0.8.4
mkl-service==2.3.0
MLFeatureSelection==0.0.9.5.1
more-itertools==8.2.0
mpmath==1.1.0
msgpack==1.0.0
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.3
nni==1.4
notebook==6.0.3
numba==0.48.0
numpy==1.18.1
oauthlib==3.0.1
olefile==0.46
opt-einsum==0+untagged.53.g6ab433b.dirty
packaging==20.1
pandas==1.0.2
pandas-profiling==2.3.0
pandocfilters==1.4.2
parso==0.6.2
partd==1.1.0
patsy==0.5.1
phik==0.9.10
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.0.0
pluggy==0.12.0
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.11.4
psutil==5.7.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
pytest==5.4.1
pytest-pylint==0.14.1
pytest-runner==5.2
python-daemon==2.2.3
python-dateutil==2.8.1
python-mimeparse==1.6.0
PythonWebHDFS==0.2.3
pytz==2019.3
PyWavelets==1.0.3
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.2.0
Requires==0.0.3
retrying==1.3.3
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
SALib==1.3.7
schema==0.7.1
scikit-image==0.15.0
scikit-learn==0.22.2.post1
scipy==1.3.1
seaborn==0.10.0
Send2Trash==1.5.0
shap==0.29.3
simplejson==3.16.0
six==1.14.0
skope-rules==1.0.0
sortedcontainers==2.1.0
SQLAlchemy==1.3.10
statsmodels==0.11.1
stopit==1.1.2
sympy==1.4
tabulate==0.8.6
tblib==1.6.0
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-quantum==0.2.0
termcolor==1.1.0
terminado==0.8.3
terminaltables==3.1.0
testpath==0.4.4
toolz==0.10.0
torch==1.4.0
torchvision==0.5.0
tornado==6.0.4
TPOT==0.10.2
tqdm==4.44.0
traitlets==4.3.3
typed-ast==1.4.1
typing-extensions==3.7.4.2
update-checker==0.16
uritemplate==3.0.1
urllib3==1.25.7
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
xlrd==1.2.0
zict==2.0.0
zipp==3.1.0
